chore(suite): token refactors
