+++ suite-common/wallet-utils/src/tokenUtils.ts
@@ -1,13 +1,14 @@
-import { NetworkSymbol } from '@suite-common/wallet-config';
+import { Network, NetworkSymbol } from '@suite-common/wallet-config';
+import { TokenInfo } from '@trezor/blockchain-link-types';
 import { parseAsset } from '@trezor/blockchain-link-utils/src/blockfrost';
 
 export const getContractAddressForNetwork = (
-    networkSymbol: NetworkSymbol,
+    networkSymbol: NetworkSymbol | (string & {}), // unknown symbols will result to lowerCase
     contractAddress: string,
 ) => {
     switch (networkSymbol) {
         case 'eth':
-            // Specyfing most common network as first case improves performance little bit
+            // Specifying most common network as first case improves performance little bit
             return contractAddress.toLowerCase();
         case 'sol':
         case 'dsol':
@@ -21,4 +22,16 @@ export const getContractAddressForNetwork = (
         default:
             return contractAddress.toLowerCase();
     }
+};
+
+export const getTokenExplorerUrl = (
+    network: Network,
+    token: Pick<TokenInfo, 'contract' | 'fingerprint'>,
+) => {
+    const explorerUrl =
+        network.networkType === 'cardano' ? network.explorer.token : network.explorer.account;
+    const contractAddress = network.networkType === 'cardano' ? token.fingerprint : token.contract;
+    const queryString = network.explorer.queryString ?? '';
+
+    return `${explorerUrl}${contractAddress}${queryString}`;
 };
